+++ test/e2e/webdriver/virtual-authenticator.ts
@@ -1,42 +1,58 @@
+import {
+  VirtualAuthenticatorOptions,
+  Transport,
+  Protocol,
+} from 'selenium-webdriver/lib/virtual_authenticator';
 import type { PasskeyRecord } from '@metamask/passkey-controller';
 import { Driver } from './driver';
 import { PlaywrightDriver } from './driver-playwright';
 
-/**
- * WebAuthn virtual authenticator helpers.
- *
- * All passkey specs are migrated to Playwright, where the virtual
- * authenticator is implemented by the Playwright driver itself via the CDP
- * WebAuthn domain (Chromium-only). The former Selenium implementation
- * (selenium-webdriver's VirtualAuthenticatorOptions) has been removed.
- *
- * The `instanceof` check also works when `driver` is the E2E_DEBUG logging
- * proxy, since Proxy forwards getPrototypeOf.
- *
- * @param driver - The driver for the current test.
- */
+type RawDriverWithVirtualAuth = {
+  addVirtualAuthenticator: (
+    options: VirtualAuthenticatorOptions,
+  ) => Promise<void>;
+  removeVirtualAuthenticator: () => Promise<void>;
+};
+
+function createPlatformAuthenticatorOptions(): VirtualAuthenticatorOptions {
+  const authOptions = new VirtualAuthenticatorOptions();
+  authOptions.setProtocol(Protocol.CTAP2);
+  authOptions.setTransport(Transport.INTERNAL);
+  authOptions.setHasResidentKey(true);
+  authOptions.setHasUserVerification(true);
+  authOptions.setIsUserVerified(true);
+  authOptions.setIsUserConsenting(true);
+  return authOptions;
+}
+
+function getRawDriver(driver: Driver): RawDriverWithVirtualAuth {
+  return driver.driver as unknown as RawDriverWithVirtualAuth;
+}
+
 export async function addVirtualAuthenticator(
   driver: Driver | PlaywrightDriver,
 ): Promise<void> {
-  if (!(driver instanceof PlaywrightDriver)) {
-    throw new Error(
-      'virtualAuthenticator is only supported on the Playwright driver. ' +
-        'Migrate the spec to Playwright (driverType: E2E_DRIVER.PLAYWRIGHT).',
-    );
+  // The Playwright driver implements the virtual authenticator itself via
+  // the CDP WebAuthn domain; the Selenium path goes through the raw
+  // WebDriver session. `instanceof` also works when `driver` is the
+  // E2E_DEBUG logging proxy, since Proxy forwards getPrototypeOf.
+  if (driver instanceof PlaywrightDriver) {
+    await driver.addVirtualAuthenticator();
+    return;
   }
-  await driver.addVirtualAuthenticator();
+  await getRawDriver(driver).addVirtualAuthenticator(
+    createPlatformAuthenticatorOptions(),
+  );
 }
 
 export async function removeVirtualAuthenticator(
   driver: Driver | PlaywrightDriver,
 ): Promise<void> {
-  if (!(driver instanceof PlaywrightDriver)) {
-    throw new Error(
-      'virtualAuthenticator is only supported on the Playwright driver. ' +
-        'Migrate the spec to Playwright (driverType: E2E_DRIVER.PLAYWRIGHT).',
-    );
+  if (driver instanceof PlaywrightDriver) {
+    await driver.removeVirtualAuthenticator();
+    return;
   }
-  await driver.removeVirtualAuthenticator();
+  await getRawDriver(driver).removeVirtualAuthenticator();
 }
 
 export const DUMMY_PASSKEY_RECORD: PasskeyRecord = {
